+++ src/Data/Core/Role/RolePolicy.php
@@ -4,10 +4,11 @@ declare(strict_types=1);
 
 namespace App\Data\Core\Role;
 
+use App\Data\AccessPolicyInterface;
 use App\Services\Core\AuthorizationService;
 use Yiisoft\User\CurrentUser;
 
-final readonly class RolePolicy
+final readonly class RolePolicy implements AccessPolicyInterface
 {
     public const GROUP = 'ROLE';
     public const ACCESS = 'ACCESS';